fix: moved manual operation tests into their own .py
